We have undone some of the work of the previous commit (not because it was a bad
direction, but because we wan't to make another change that will mean that the
previous work would need to be change anyway.) We have also change the code to
suppress some warnings (although there are still quite a few.) We have also
started to add the changes required to implement the aformentioned change.
We are going to add the abilitry to scroll in the y axis as well as the x axis.
However to avoid large files and RAM usage will will have to emply a format for
our levels where each level is made up of blocks (chunks if you will) that are
the size of the window.

+++ game/src/levelRules.cpp
@@ -199,12 +199,12 @@ void rules::movePlayerRightWhenInteractingWithInnerMargin
     { // We are to the left of the inner right margin.
       gamePlayer->updatePosRel(input);
     }
-  else if((position + maxyx.x) < backgroundLength)
+  else if(size_t(position + maxyx.x) < backgroundLength)
     { // There is still background left to spare.
       gamePlayer->updateDirection(input);
       position++;	// Move background.
     }
-  else if((position + maxyx.x) == backgroundLength)
+  else if(size_t(position + maxyx.x) == backgroundLength)
     { /* No background left, so move the player to the right
 	 edge of the background. */
       gamePlayer->updatePosRel(input);
